correct controllers

+++ shop-project/src/main/java/com/online/shop/service/impl/CategoryServiceImpl.java
@@ -1,5 +1,8 @@
 package com.online.shop.service.impl;
 
+import com.online.shop.entity.User;
+import com.online.shop.exceptions.InvalidEntityDataException;
+import com.online.shop.exceptions.NotFoundEntityException;
 import com.online.shop.service.CategoryService;
 import com.online.shop.dto.CategoryDto;
 import com.online.shop.entity.Category;
@@ -11,6 +14,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Slf4j
@@ -29,16 +33,20 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public Long insert(CategoryDto entityDto) {
+        log.info("Executing insert method in CategoryServiceImpl with DTO: {}", entityDto);
         if (entityDto == null) {
             log.error("CategoryDto is null in insert method");
             throw new IllegalArgumentException("CategoryDto cannot be null");
         }
-        log.info("Executing insert method in CategoryServiceImpl with DTO: {}", entityDto);
+        if (entityDto.getName() == null || entityDto.getDescription() == null) {
+            throw new InvalidEntityDataException("Data are required");
+        }
         return categoryDao.insert(modelMapper.map(entityDto, Category.class));
     }
 
     @Override
     public CategoryDto update(Long id, CategoryDto entityDto) {
+        log.info("Executing update method in CategoryServiceImpl for ID: {} with DTO: {}", id, entityDto);
         if (id == null) {
             log.error("ID is null in update method");
             throw new IllegalArgumentException("ID cannot be null");
@@ -47,17 +55,25 @@ public class CategoryServiceImpl implements CategoryService {
             log.error("CategoryDto is null in update method");
             throw new IllegalArgumentException("CategoryDto cannot be null");
         }
-        log.info("Executing update method in CategoryServiceImpl for ID: {} with DTO: {}", id, entityDto);
-        return modelMapper.map(categoryDao.update(id, modelMapper.map(entityDto, Category.class)), CategoryDto.class);
+        Optional<Category> updatedCategory = categoryDao.update(id, modelMapper.map(entityDto, Category.class));
+        if (!updatedCategory.isPresent()) {
+            throw new NotFoundEntityException("Category not found");
+        }
+        return modelMapper.map(updatedCategory, CategoryDto.class);
     }
 
     @Override
     public CategoryDto findById(Long id) {
+        log.info("Executing findById method in CategoryServiceImpl for ID: {}", id);
         if (id == null) {
             log.error("ID is null in findById method");
             throw new IllegalArgumentException("ID cannot be null");
         }
-        log.info("Executing findById method in CategoryServiceImpl for ID: {}", id);
+        Optional<Category> foundCategory = categoryDao.getById(id);
+        if (!foundCategory.isPresent()) {
+            log.error("Category not found");
+            throw new NotFoundEntityException("Category not found");
+        }
         return modelMapper.map(categoryDao.getById(id), CategoryDto.class);
     }
 
@@ -71,11 +87,16 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public Boolean delete(Long id) {
+        log.info("Executing delete method in CategoryServiceImpl for ID: {}", id);
         if (id == null) {
             log.error("ID is null in delete method");
             throw new IllegalArgumentException("ID cannot be null");
         }
-        log.info("Executing delete method in CategoryServiceImpl for ID: {}", id);
+        Boolean deleted = categoryDao.delete(id);
+        if (!deleted) {
+            log.error("Category not found");
+            throw new NotFoundEntityException("User not found");
+        }
         return categoryDao.delete(id);
     }
